update aplikasi tanggal 6/19/23 setelah sidang hari rabu

+++ application/models/Super_Admin.php
@@ -1,5 +1,5 @@
 <?php
-class Admin_model extends CI_Model {
+class Super_Admin extends CI_Model {
   
     public function get_data_with_join() {
       $this->db->select('materi_quran.*, surah.surah, surah.id_surah');
@@ -7,7 +7,6 @@ class Admin_model extends CI_Model {
       $this->db->join('surah', 'surah.id_surah = materi_quran.id_surah');
       $query = $this->db->get();
       return $query->result_array();
-      
   }
   public function get_data_by_id($id) {
     $this->db->where('id_materi', $id);
@@ -21,15 +20,20 @@ class Admin_model extends CI_Model {
     $this->db->update('materi_quran', $data);
     
   }
-  
-  public function delete_data($id) {
-    $this->db->where('id_materi', $id);
-    $this->db->delete('materi_quran');
+  public function delate_data($table, $id) {
+      if ($table === 'materi_quran') {
+          $this->db->where('id_materi', $id);
+          $this->db->delete('materi_quran');
+      }
+      if ($table === 'admin') {
+          $this->db->where('id', $id);
+          $this->db->delete('admin');
+      }
   }
-  
-  public function createData($data)  {
-        return $this->db->insert('materi_quran', $data);
-    }
+
+ public function createData($table, $data) {
+    return $this->db->insert($table, $data);
+}
 
   public function getSurahNames() {
         $this->db->select('id_surah, surah');
@@ -37,9 +41,18 @@ class Admin_model extends CI_Model {
         return $query->result_array();
     }
   public function getdataAdmin(){
-  return $this->db->get('admin')->result_array();
-
+      return $this->db->get('admin')->result_array();
   }
+  public function get_data_byuser($user) {
+      $this->db->select('materi_quran.*, surah.surah, surah.id_surah');
+      $this->db->from('materi_quran');
+      $this->db->join('surah', 'surah.id_surah = materi_quran.id_surah');
+     $this->db->where('materi_quran.oleh', $user);
+    $query = $this->db->get();
+
+    return $query->result_array();
+}
+
 }
 
 ?>
